requirements.txt
discord.py[voice]
requests
lavalink
Wavelink
octorest